backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.1
pydantic>=2.10.0
requests>=2.32.0
python-dotenv>=1.0.1
PyJWT>=2.8.0
qrcode[pil]>=7.4.2
Pillow>=10.4.0
python-jose[cryptography]>=3.3.0
redis>=5.0.0
supabase>=2.0.0
pytesseract>=0.3.10
pdfplumber>=0.11.0
pdf2image>=1.17.0
python-multipart>=0.0.9
